pyproject.toml
[project]
name = "capstoneprojectdatasense"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "azure-identity==1.25.3",
    "beautifulsoup4==4.14.3",
    "fastapi==0.122.0",
    "gunicorn==23.0.0",
    "httpx==0.28.1",
    "langchain-core==1.3.0",
    "langchain-openai==1.2.0",
    "langgraph==1.1.8",
    "neo4j==6.1.0",
    "openai==2.31.0",
    "openpyxl==3.1.5",
    "pageindex==0.2.8",
    "pydantic==2.12.5",
    "psycopg[binary]==3.3.3",
    "PyMuPDF==1.26.4",
    "python-dotenv==1.1.0",
    "python-jose[cryptography]==3.5.0",
    "requests==2.32.5",
    "uvicorn==0.38.0",
]

[tool.pytest.ini_options]
testpaths = ["backend/tests", "tests"]
pythonpath = [".", "backend"]
